+++ packages/create-theme/public/template/docs/.vitepress/blog-theme.ts
@@ -66,32 +66,35 @@ const blogTheme = getThemeConfig({
   popover: {
     title: '公告',
     body: [
-      {
-        type: 'text',
-        content:
-          'QQ交流群：681489336 🎉🎉',
-      },
-      {
-        type: 'text',
-        content:
-          '👇公众号👇---👇 微信 👇',
-      },
+      { type: 'text', content: '👇公众号👇---👇 微信 👇' },
       {
         type: 'image',
-        src: 'https://img.cdn.sugarat.top/mdImg/MTYxNTAxODc2NTIxMA==615018765210',
+        src: 'https://img.cdn.sugarat.top/mdImg/MTYxNTAxODc2NTIxMA==615018765210'
       },
       {
         type: 'text',
-        content:
-          '欢迎大家加群&私信交流',
+        content: '欢迎大家加群&私信交流'
+      },
+      {
+        type: 'text',
+        content: '文章首/文尾有群二维码',
+        style: 'padding-top:0'
       },
       {
         type: 'button',
-        content: '博客',
-        link: 'https://sugarat.top',
+        content: '作者博客',
+        link: 'https://sugarat.top'
       },
+      {
+        type: 'button',
+        content: '加群交流',
+        props: {
+          type: 'success'
+        },
+        link: 'https://theme.sugarat.top/group.html',
+      }
     ],
-    duration: 0,
+    duration: 0
   },
 })