+++ packages/theme/src/composables/config/index.ts
@@ -1,7 +1,9 @@
 /* eslint-disable ts/no-namespace */
 import type { ElButton } from 'element-plus'
 import type { DefaultTheme } from 'vitepress'
-import type { FeedOptions } from 'feed'
+import type { RSSOptions } from 'vitepress-plugin-rss'
+
+type RSSPluginOptions = RSSOptions
 
 // TODO: 重构 lint 问题
 export declare namespace BlogPopover {
@@ -304,40 +306,7 @@ export namespace Theme {
     footer?: Footer
   }
 
-  export type RSSOptions = Omit<FeedOptions, 'id'> & {
-    id?: string
-    /**
-     * 你的站点地址
-     * @example 'https://sugarat.top'
-     */
-    baseUrl: string
-    /**
-     * 线上访问的RSS地址
-     * @default
-     * @example https://sugarat.top/feed.rss
-     * ```ts
-     * `${baseUrl + VPConfig.site.base + (filename || 'feed.rss'}`
-     * ```
-     */
-    url?: string
-    /**
-     * 输出的RSS文件名
-     * @default 'feed.rss'
-     */
-    filename?: string
-    /**
-     * RSS的图标展示
-     * @default true
-     */
-    icon?: boolean
-    /**
-     * 限制输出文件包含的文章数量
-     * @default 0
-     * @description (0 不限制；> 1 会按照日期排序对输出内容进行调整)
-     */
-    limit?: number
-    ariaLabel?: string
-  }
+  export type RSSOptions = RSSPluginOptions
 
   export interface Footer {
     /**