+++ packages/theme/src/utils/node/vitePlugins.ts
@@ -6,8 +6,8 @@ import {
   chineseSearchOptimize,
   pagefindPlugin
 } from 'vitepress-plugin-pagefind'
+import { RssPlugin } from 'vitepress-plugin-rss'
 import type { Theme } from '../../composables/config/index'
-import { genFeed } from './genFeed'
 
 export function getVitePlugins(cfg?: Partial<Theme.BlogConfig>) {
   const plugins: any[] = []
@@ -24,6 +24,17 @@ export function getVitePlugins(cfg?: Partial<Theme.BlogConfig>) {
       pagefindPlugin({ ...ops, customSearchQuery: chineseSearchOptimize })
     )
   }
+
+  // 内置支持Mermaid
+  if (cfg?.mermaid !== false) {
+    const { MermaidPlugin } = require('vitepress-plugin-mermaid')
+    plugins.push(MermaidPlugin(cfg?.mermaid === true ? {} : (cfg?.mermaid ?? {})))
+  }
+
+  // 内置支持RSS
+  if (cfg?.RSS) {
+    plugins.push(RssPlugin(cfg.RSS))
+  }
   // 未来移除使用
   // if (cfg && cfg.search !== undefined) {
   //   console.log(
@@ -31,7 +42,6 @@ export function getVitePlugins(cfg?: Partial<Theme.BlogConfig>) {
   //   )
   // }
 
-  buildEndFn.push(genFeed)
   return plugins
 }