+++ packages/blogpress/CHANGELOG.md
@@ -3,6 +3,13 @@ publish: false
 ---
 
 # blogpress
+
+## 2.0.12
+### Patch Changes
+
+- Updated dependencies
+  - @sugarat/theme@0.2.2
+
 ## 2.0.11
 
 ### Patch Changes