+++ packages/create-theme/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sugarat/create-theme
 
+## 0.0.20
+
+### Patch Changes
+
+- feat: 简化模板配置
+
+## 0.0.19
+
+### Patch Changes
+
+- feat: 主题色定制的wiki更新
+
 ## 0.0.18
 
 ### Patch Changes