+++ packages/theme/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sugarat/theme
 
+## 0.2.2
+
+### Patch Changes
+
+- feat: 默认开启 tabs 支持
+- chore: 优化 Mermaid 配置的生成逻辑
+- chore: 内置RSS替换为 [vitepress-plugin-rss](https://www.npmjs.com/package/vitepress-plugin-tabs)
+- chore: 架构优化
+  - 便于独立在VitePress使用
+  - 无用依赖清理
+  - 主题无关配置屏蔽
+  - 无用依赖移除
+
 ## 0.2.1
 
 ### Patch Changes